Update existing py-tess package
